Adding appointment types
Each qnaire can now have appointment types added to it.  Appointments
may then be labelled with these types such that the appointment list
pull widget will only show appointments of the types specified, or appointments
which are not marked as "special" if no types are provided when asking
for the appointment list.